[TC] Regenerated demo for tc_tc_compare_mode_sam_v71_xult
MH3-67830

+++ apps/tc/tc_compare_mode/firmware/src/config/sam_v71_xult/peripheral/tc/plib_tc0.c
@@ -112,7 +112,7 @@ void TC0_CH0_ComparePeriodSet (uint16_t period)
 /* Read the period value */
 uint16_t TC0_CH0_ComparePeriodGet (void)
 {
-    return TC0_REGS->TC_CHANNEL[0].TC_RC;
+    return (uint16_t)TC0_REGS->TC_CHANNEL[0].TC_RC;
 }
 
 /* Set the compare A value */
